+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 12/02/2021
+
+- Remove dependency on CT's `authenticated` functionality
+
 ## 10/02/2021
 
 - Add error handler and message for the "too_many_buckets" error when using too many "group by" values.